Improved updating
There is no need to `clear` and `refresh` unless `needs_update` is set.

If the clock is updated for the new time, then the region reguires updating is still the same. But if `needs_update` is True, then some option is changed, a screen clear is the best option to ensure no artifacts. Although color change doesn't need to whole screen cleared.